+++ clemcore/playpen/envs/__init__.py
@@ -1,30 +1,5 @@
 import abc
-from contextlib import contextmanager
 from typing import List, Tuple, Callable, Union, Dict
-
-from clemcore.backends import Model
-from clemcore.clemgame import GameSpec, benchmark
-
-from clemcore.playpen.envs.game_env import GameEnv
-from clemcore.playpen.envs.tree_env import GameTreeEnv
-
-
-@contextmanager
-def make_env(game_spec: GameSpec, players: List[Model],
-             instances_name: str = None, shuffle_instances: bool = False):
-    with benchmark.load_from_spec(game_spec, do_setup=True, instances_name=instances_name) as game:
-        task_iterator = game.create_game_instance_iterator(shuffle_instances)
-        yield GameEnv(game, players, task_iterator)
-
-
-@contextmanager
-def make_tree_env(game_spec: GameSpec, players: List[Model],
-                  instances_name: str = None, shuffle_instances: bool = False,
-                  branching_factor: int = 2, pruning_fn=lambda candidates: candidates):
-    with benchmark.load_from_spec(game_spec, do_setup=True, instances_name=instances_name) as game:
-        assert branching_factor > 1, "The branching factor must be greater than one"
-        task_iterator = game.create_game_instance_iterator(shuffle_instances)
-        yield GameTreeEnv(game, players, task_iterator, branching_factor=branching_factor, pruning_fn=pruning_fn)
 
 
 class PlayPenEnv(abc.ABC):
